feat(picking-packing): agregar nuevas rutas y componentes para la gestión de picking y packing. Se implementan formularios para iniciar nuevos procesos de picking y packing, se añaden modelos para las órdenes y se optimizan los componentes existentes. Además, se mejora la lógica de carga de datos y se actualizan los estilos para una mejor experiencia de usuario.

+++ src/app/components/ventas/carrito/carrito.component.ts
@@ -176,7 +176,7 @@ export class CarritoComponent implements OnInit {
         }
 
         this.valorDescuento = 0;
-        this.porcentajeDescuento = parseInt(value[0]?.valor) || 0;
+        this.porcentajeDescuento = parseFloat(value[0]?.valor) || 0;
         this.pedido.porceDescuento = this.porcentajeDescuento;
         this.valorDescuento = (this.getTotalProductPriceInCart() * this.porcentajeDescuento) / 100;
       },
